requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
python-multipart>=0.0.6
httpx>=0.26.0
gunicorn>=21.2.0
tenacity>=8.2.3

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
ruff>=0.1.14

